poll editor done
