+++ packages/commonwealth/client/scripts/views/pages/view_proposal/poll_editor_card.tsx
@@ -3,21 +3,23 @@
 import m from 'mithril';
 
 import 'pages/view_proposal/poll_editor_card.scss';
+
+import app from 'state';
 import { Thread } from 'models';
 import { CWCard } from '../../components/component_kit/cw_card';
 import { CWButton } from '../../components/component_kit/cw_button';
 import { CWText } from '../../components/component_kit/cw_text';
+import { PollEditorModal } from '../../modals/poll_editor_modal';
 
 export class PollEditorCard
   implements
     m.ClassComponent<{
       proposal: Thread;
       proposalAlreadyHasPolling: boolean;
-      openPollEditor: () => void;
     }>
 {
   view(vnode) {
-    const { proposal, proposalAlreadyHasPolling, openPollEditor } = vnode.attrs;
+    const { proposal, proposalAlreadyHasPolling } = vnode.attrs;
 
     return (
       <CWCard className="PollEditorCard">
@@ -27,12 +29,15 @@ export class PollEditorCard
         </CWText>
         <CWButton
           disabled={!!proposal.offchainVotingEndsAt}
-          label={
-            proposal.votingEndTime ? 'Polling enabled' : 'Create poll'
-          }
+          label={proposal.votingEndTime ? 'Polling enabled' : 'Create poll'}
           onclick={(e) => {
             e.preventDefault();
-            openPollEditor();
+            app.modals.create({
+              modal: PollEditorModal,
+              data: {
+                thread: proposal,
+              },
+            });
           }}
         />
       </CWCard>
