--- packages/commonwealth/client/styles/components/poll_editor.scss
@@ -1,72 +1,0 @@
-@import 'client/styles/shared';
-
-.PollEditorDialog {
-  .cui-dialog-body {
-    padding: 24px;
-
-    .cui-switch {
-      color: $neutral-800;
-      margin-bottom: 16px;
-    }
-
-    h4 {
-      margin: 8px 0;
-    }
-
-    h4.disabled {
-      color: $neutral-500;
-    }
-
-    .poll-editor-options-question {
-      margin-bottom: 16px;
-    }
-
-    .poll-editor-options {
-      .poll-editor-choice-buttons {
-        margin: 4px 0;
-      }
-
-      .poll-editor-choice {
-        margin-bottom: 4px;
-      }
-
-      .cui-input,
-      .cui-button {
-        margin-bottom: 8px;
-      }
-    }
-
-    .poll-duration-copy {
-      color: $neutral-500;
-      margin: 16px 0px;
-    }
-
-    .poll-editor-duration {
-      align-items: center;
-      display: flex;
-      justify-content: space-between;
-
-      .cui-switch {
-        color: $neutral-500;
-        margin-bottom: 0;
-        min-width: max-content;
-      }
-    }
-  }
-}
-
-.custom-duration-items {
-  .cui-popover-content {
-    padding: 8px;
-    width: 104px;
-
-    .duration-item {
-      cursor: pointer;
-      padding: 4px 8px;
-
-      &:hover {
-        background-color: $primary-400;
-      }
-    }
-  }
-}
